Move compose logic to lib package
