+++ lib/compose/README.md
